vizsgaremek/requirements.txt
mysql_connector_repackaged==0.3.1
pytest==7.1.1
selenium==4.2.0
webdriver_manager==3.5.4
mysql-connector-python==8.0.29
